@@ -1,6 +1,6 @@
 class Users::RegistrationsController < Devise::RegistrationsController
- before_filter :configure_sign_up_params, only: [:create]
- before_filter :configure_account_update_params, only: [:update]
+ before_action :configure_sign_up_params, only: [:create]
+ before_action :configure_account_update_params, only: [:update]
 
   # GET /resource/sign_up
   # def new
@@ -46,10 +46,10 @@ class Users::RegistrationsController < Devise::RegistrationsController
   # If you have extra params to permit, append them to the sanitizer.
    def configure_account_update_params
      devise_parameter_sanitizer.for(:account_update) { |u| u.permit(
-                                                      :username, 
-                                                      :email, 
-                                                      :password, 
-                                                      :current_password, 
+                                                      :username,
+                                                      :email,
+                                                      :password,
+                                                      :current_password,
                                                       :password_confirmation) }
    end
 
