@@ -1,5 +1,5 @@
 class Users::SessionsController < Devise::SessionsController
- before_filter :configure_sign_in_params, only: [:create]
+ before_action :configure_sign_in_params, only: [:create]
 
   # GET /resource/sign_in
   # def new
@@ -20,6 +20,6 @@ class Users::SessionsController < Devise::SessionsController
 
   # If you have extra params to permit, append them to the sanitizer.
    def configure_sign_in_params
-     devise_parameter_sanitizer.for(:sign_in) { |u| u.permit(:email, :password) }
+     devise_parameter_sanitizer.permit(:sign_in) { |u| u.permit(:email, :password) }
    end
 end
